break up export-to-notebook.spec test
